requirements.txt
astroid==2.3.3
autopep8==1.5
Click==7.0
entrypoints==0.3
flake8==3.7.9
Flask==1.1.1
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.1
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
pycodestyle==2.5.0
pyflakes==2.1.1
pymongo==3.10.1
python-dotenv==0.12.0
six==1.14.0
typed-ast==1.4.1
Werkzeug==1.0.0
wrapt==1.11.2
